+++ nequip/data/__init__.py
@@ -9,7 +9,7 @@ from .AtomicData import (
     _LONG_FIELDS,
 )
 from .dataset import AtomicDataset, AtomicInMemoryDataset, NpzDataset, ASEDataset
-from .dataloader import DataLoader, Collater
+from .dataloader import DataLoader, Collater, PartialSampler
 from ._build import dataset_from_config
 from ._test_data import EMTTestDataset
 
@@ -24,6 +24,7 @@ __all__ = [
     ASEDataset,
     DataLoader,
     Collater,
+    PartialSampler,
     dataset_from_config,
     _NODE_FIELDS,
     _EDGE_FIELDS,