+++ nequip/data/dataset.py
@@ -6,6 +6,7 @@ import functools
 import itertools
 import yaml
 import hashlib
+import math
 from os.path import dirname, basename, abspath
 from typing import Tuple, Dict, Any, List, Callable, Union, Optional, Sequence
 
@@ -40,12 +41,14 @@ class AtomicDataset(Dataset):
     """The base class for all NequIP datasets."""
 
     root: str
+    dtype: torch.dtype
 
     def __init__(
         self,
         root: str,
         type_mapper: Optional[TypeMapper] = None,
     ):
+        self.dtype = torch.get_default_dtype()
         super().__init__(root=root, transform=type_mapper)
 
     def statistics(
@@ -79,7 +82,7 @@ class AtomicDataset(Dataset):
             if k not in IGNORE_KEYS and hasattr(self, k)
         }
         # Add other relevant metadata:
-        params["dtype"] = str(torch.get_default_dtype())
+        params["dtype"] = str(self.dtype)
         params["nequip_version"] = nequip.__version__
         return params
 
@@ -411,9 +414,7 @@ class AtomicInMemoryDataset(AtomicDataset):
             if callable(field):
                 # make a joined thing? so it includes fixed fields
                 arr, arr_is_per = field(data_transformed)
-                arr = arr.to(
-                    torch.get_default_dtype()
-                )  # all statistics must be on floating
+                arr = arr.to(self.dtype)  # all statistics must be on floating
                 assert arr_is_per in ("node", "graph", "edge")
             else:
                 if field not in all_keys:
@@ -442,7 +443,7 @@ class AtomicInMemoryDataset(AtomicDataset):
                 )
             if not isinstance(arr, torch.Tensor):
                 if np.issubdtype(arr.dtype, np.floating):
-                    arr = torch.as_tensor(arr, dtype=torch.get_default_dtype())
+                    arr = torch.as_tensor(arr, dtype=self.dtype)
                 else:
                     arr = torch.as_tensor(arr)
             if arr_is_per == "node":
@@ -542,8 +543,8 @@ class AtomicInMemoryDataset(AtomicDataset):
                 f"{ana_mode} for per-atom analysis is not implemented"
             )
 
-    @staticmethod
     def _per_species_statistics(
+        self,
         ana_mode: str,
         arr: torch.Tensor,
         arr_is_per: str,
@@ -569,12 +570,12 @@ class AtomicInMemoryDataset(AtomicDataset):
                     f"{ana_mode} for per species analysis is not implemented for shape {arr.shape}"
                 )
 
-            N = N.type(torch.get_default_dtype())
+            N = N.type(self.dtype)
 
             return solver(N, arr, **algorithm_kwargs)
 
         elif arr_is_per == "node":
-            arr = arr.type(torch.get_default_dtype())
+            arr = arr.type(self.dtype)
 
             if ana_mode == "mean_std":
                 mean = scatter_mean(arr, atom_types, dim=0)
@@ -594,6 +595,58 @@ class AtomicInMemoryDataset(AtomicDataset):
 
         else:
             raise NotImplementedError
+
+    def rdf(
+        self, bin_width: float, stride: int = 1
+    ) -> Dict[Tuple[int, int], Tuple[np.ndarray, np.ndarray]]:
+        """Compute the pairwise RDFs of the dataset.
+
+        Args:
+            bin_width: width of the histogram bin in distance units
+            stride: stride of data to include
+
+        Returns:
+            dictionary mapping `(type1, type2)` to tuples of `(hist, bin_edges)` in the style of `np.histogram`.
+        """
+        graph_selector, node_selector, edge_selector = self._selectors(stride=stride)
+
+        data = AtomicData.to_AtomicDataDict(self.data)
+        data = AtomicDataDict.with_edge_vectors(data, with_lengths=True)
+
+        results = {}
+
+        types = self.type_mapper(data)[AtomicDataDict.ATOM_TYPE_KEY]
+
+        edge_types = torch.index_select(
+            types, 0, data[AtomicDataDict.EDGE_INDEX_KEY].reshape(-1)
+        ).view(2, -1)
+        types_center = edge_types[0].numpy()
+        types_neigh = edge_types[1].numpy()
+
+        r_max: float = self.AtomicData_options["r_max"]
+        # + 1 to always have a zero bin at the end
+        n_bins: int = int(math.ceil(r_max / bin_width)) + 1
+        # +1 since these are bin_edges including rightmost
+        bins = bin_width * np.arange(n_bins + 1)
+
+        for type1, type2 in itertools.combinations(
+            range(self.type_mapper.num_types), 2
+        ):
+            # Try to do as much of this as possible in-place
+            mask = types_center == type1
+            np.logical_and(mask, types_neigh == type2, out=mask)
+            np.logical_and(mask, edge_selector, out=mask)
+            mask = mask.astype(np.int32)
+            results[(type1, type2)] = np.histogram(
+                data[AtomicDataDict.EDGE_LENGTH_KEY],
+                weights=mask,
+                bins=bins,
+                density=True,
+            )
+            # RDF is symmetric
+            results[(type2, type1)] = results[(type1, type2)]
+
+        return results
 
 
 class NpzDataset(AtomicInMemoryDataset):