+++ CHANGELOG.md
@@ -7,11 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Most recent change on the bottom.
 
 ## Unreleased - 0.6.0
+### Added
+- [Breaking] Refactor overall model logic into `GraphModel` top-level module
+- [Breaking] Added `model_dtype`
+- `BATCH_PTR_KEY` in `AtomicDataDict`
+- `AtomicInMemoryDataset.rdf()`
+
 ### Changed
 - Always require explicit `seed`
 - [Breaking] Set `dataset_seed` to `seed` if it is not explicitly provided
 - Don't log as often by default
 - [Breaking] Default nonlinearities are `silu` (`e`) and `tanh` (`o`)
+- Will not reproduce previous versions' data shuffling order (for all practical purposes this does not matter, the `shuffle` option is unchanged)
+- [Breaking] `default_dtype` defaults to `float64` (`model_dtype` default `float32`)
+- `nequip-benchmark` now only uses `--n-data` frames to build the model
 
 ### Fixed
 - Work with `wandb>=0.13.8`